refactor(ui): deprecating some components
